requirements.txt
nose
msrestazure>=0.4.0,<0.5.0
python-dateutil